pyproject.toml
[tool.poetry]
name = "llm-chat-engine"
version = "0.1.0"
description = ""
authors = ["George Fatouros <gfatouros@innov-acts.com>"]
readme = "README.md"
packages = [{include = "llm_chat_engine"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
llama-index = "^0.10.41"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
